images/instrumentation/python/requirements.txt
# Modelled after
# https://github.com/open-telemetry/opentelemetry-operator/blob/v0.143.0/autoinstrumentation/python/requirements.txt

opentelemetry-distro==0.60b1

urllib3 <2.6.4
opentelemetry-exporter-otlp-proto-http==1.39.1
opentelemetry-exporter-prometheus==0.60b1

opentelemetry-propagator-b3==1.39.1
opentelemetry-propagator-jaeger==1.39.1
opentelemetry-propagator-aws-xray==1.0.2

opentelemetry-instrumentation==0.60b1
opentelemetry-propagator-ot-trace==0.60b1

opentelemetry-instrumentation-aio-pika==0.60b1
opentelemetry-instrumentation-aiohttp-client==0.60b1
opentelemetry-instrumentation-aiohttp-server==0.60b1
opentelemetry-instrumentation-aiokafka==0.60b1
opentelemetry-instrumentation-aiopg==0.60b1
opentelemetry-instrumentation-asgi==0.60b1
opentelemetry-instrumentation-asyncio==0.60b1
opentelemetry-instrumentation-asyncpg==0.60b1
opentelemetry-instrumentation-aws-lambda==0.60b1
opentelemetry-instrumentation-boto==0.60b1
opentelemetry-instrumentation-boto3sqs==0.60b1
opentelemetry-instrumentation-botocore==0.60b1
opentelemetry-instrumentation-cassandra==0.60b1
opentelemetry-instrumentation-celery==0.60b1
opentelemetry-instrumentation-click==0.60b1
opentelemetry-instrumentation-confluent-kafka==0.60b1
opentelemetry-instrumentation-dbapi==0.60b1
opentelemetry-instrumentation-django==0.60b1
opentelemetry-instrumentation-elasticsearch==0.60b1
opentelemetry-instrumentation-falcon==0.60b1
opentelemetry-instrumentation-fastapi==0.60b1
opentelemetry-instrumentation-flask==0.60b1
opentelemetry-instrumentation-grpc==0.60b1
opentelemetry-instrumentation-httpx==0.60b1
opentelemetry-instrumentation-jinja2==0.60b1
opentelemetry-instrumentation-kafka-python==0.60b1
opentelemetry-instrumentation-logging==0.60b1
opentelemetry-instrumentation-mysql==0.60b1
opentelemetry-instrumentation-mysqlclient==0.60b1
opentelemetry-instrumentation-pika==0.60b1
opentelemetry-instrumentation-psycopg==0.60b1
opentelemetry-instrumentation-psycopg2==0.60b1
opentelemetry-instrumentation-pymemcache==0.60b1
opentelemetry-instrumentation-pymongo==0.60b1
opentelemetry-instrumentation-pymysql==0.60b1
opentelemetry-instrumentation-pyramid==0.60b1
opentelemetry-instrumentation-redis==0.60b1
opentelemetry-instrumentation-remoulade==0.60b1
opentelemetry-instrumentation-requests==0.60b1
opentelemetry-instrumentation-sqlalchemy==0.60b1
opentelemetry-instrumentation-sqlite3==0.60b1
opentelemetry-instrumentation-starlette==0.60b1
opentelemetry-instrumentation-system-metrics==0.60b1
opentelemetry-instrumentation-threading==0.60b1
opentelemetry-instrumentation-tornado==0.60b1
opentelemetry-instrumentation-tortoiseorm==0.60b1
opentelemetry-instrumentation-urllib==0.60b1
opentelemetry-instrumentation-urllib3==0.60b1
opentelemetry-instrumentation-wsgi==0.60b1
